feat: Utilizado crud de serviços na página de cadastro de serviços
