+++ src/serviços.html
@@ -144,7 +144,6 @@
             id="imagem"
             name="foto"
             accept="image/*"
-            onchange="previewPicture()"
             required />
         </div>
 
@@ -245,6 +244,6 @@
       </div>
     </footer>
 
-    <script src="static/js/serviços.js"></script>
+    <script type="module" src="static/js/serviços.mjs"></script>
   </body>
 </html>
